Review from @boocmp: move the tests
